feat: add generateMetadata and generateStaticParams to project detail pages
Convert project detail page from client component to server component,
extracting interactive rendering into a separate client component. Enables
static generation of all project pages at build time and proper metadata
(title + description) for SEO and social sharing.

Closes #16
